fix: PR merge for CI systems using Checks API
Issue
-----

Github Actions Workflows use the Checks API, as opposed to the Statues API, to
post the status of a Workflow.

Current behavior of using the `/repos/{owner}/{repo}/statuses/{sha}` API to check
if a PR is ready to be merged naturally only works for CI systems that use
the Statues API (Jenkins?). One hack for GH Actions Workflows is to have a step
that posts to the Statuses API, but that gets pretty tiresome after the second
workflow.

Right now octopilot also needs read-only repository **admin** permissions to be able
to `GET /repos/{owner}/{repo}/branches/{branch}/protection/required_status_checks`
for a list of required status checks configured as part of the branch protection rules.

Fix
---

Use the PR `mergeable_state` field to determine when a PR is ready to be merged.
This field is not documented very well, but a value of `clean` *should* mean
all branch protection rules are satisfied.
This should cover both the Statues API & Checks API.
Added bonus is any other branch protection rules (e.g. required approvals).
Also, reduces the scope of the permissions needed.