feat: allow merging PRs with Github's auto-merge
This adds support for Github's auto-merge feature:

- https://docs.github.com/en/repositories/configuring-branches-and-merges-in-your-repository/configuring-pull-request-merges/managing-auto-merge-for-pull-requests-in-your-repository
- https://docs.github.com/en/pull-requests/collaborating-with-pull-requests/incorporating-changes-from-a-pull-request/automatically-merging-a-pull-request

This can only be done via the GraphQL API, so I've had to vendor another library:

https://docs.github.com/en/graphql/reference/mutations#enablepullrequestautomerge

To maintain backwards compatibilty, the exisiting method of merging a PR is left
intact and as the default.

Two new flags are introduced, `--pr-merge-auto` & `--pr-merge-auto-wait`.

`--pr-merge-auto` enables "auto-merge" for the PR (via a GraphQL mutation).
This does not wait for the PR to actually be merged. That's left up to Github
once all branch protection rules pass.

`--pr-merge-auto-wait` can optionally wait until the PR is actually merged by
Github.

Both of these flags are also exposed on a per repository basis using the
`mergeauto` & `mergeautowait` options.